requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
billiard==3.3.0.23
boto3==1.15.4
botocore==1.18.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-celery==3.3.1
django-celery-results==1.2.1
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-cron==0.5.1
django-filter==2.3.0
django-multiselectfield==0.1.12
django-rename-app==0.1.2
django-rest-auth==0.9.5
django-storages==1.10.1
django-timezone-field==4.0
djangorestframework==3.11.1
future==0.18.2
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==3.0.37
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
twilio==6.45.4
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
